fix rendering on map
